Refactored panel management
